requirements.txt
streamlit==1.31.1
pandas==2.2.0
openai==1.12.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
folium==0.15.1
streamlit-folium==0.17.0
geopy==2.4.1